* new MRC: 32bit values are now used to represent file access modes
git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@90 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534

+++ servers/src/org/xtreemfs/new_mrc/ac/FileAccessManager.java
@@ -116,7 +116,7 @@ public class FileAccessManager {
         return getVolumeFileAccessPolicy(volumeId).translateAccessMode(accessMode);
     }
     
-    public short getPosixAccessMode(StorageManager sMan, FileMetadata file, String userId,
+    public int getPosixAccessMode(StorageManager sMan, FileMetadata file, String userId,
         List<String> groupIds) throws MRCException {
         return getVolumeFileAccessPolicy(sMan.getVolumeId()).getPosixAccessRights(sMan, file,
             userId, groupIds);
